fn_80136E00: y int drops wrap-path y1=y+2 truncation mask (98.0->98.3)
In the line-wrap path y reused the just-stored debugPrintYpos and re-masked at
y1=y+2; target reuses the already-masked value. Retype y to int (lhz zero-extend)
so the +2 needs no clrlwi, matching target.

+++ src/main/dll/dll_80136a40.c
@@ -1020,7 +1020,7 @@ int fn_80136E00(int p1, u8* p)
     u8 c;
     int w;
     int x2;
-    u16 y;
+    int y;
     int y0;
     int y1;
     u16 x0;
